.history/requirement_20250428125548.txt
flask==2.1.2
mysql-connector-python==8.0.27
scikit-learn==1.0.2
numpy==1.21.2
